server/pyproject.toml
[project]
name = "dockit-server"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "openai>=1.12.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.0.0",
    "prometheus-client>=0.20.0",
    "python-dotenv>=1.0.0",
]
